requirements.txt
biopython==1.70
decorator==4.2.1
numpy==1.14.0
pycodestyle==2.3.1
regex==2018.2.21
